Move logic from main into lib in cda-main.
